listen by ms

+++ monitor/monitor_binance_future.py
@@ -52,7 +52,7 @@ async def main(symbol, delay, output_format, depth):
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description="Monitor the spot price of a trading pair on Binance.")
     parser.add_argument("symbol", help="Trading pair symbol (e.g., BTC/USDT).")
-    parser.add_argument("delay", type=int, help="Time delay between price fetches in seconds.")
+    parser.add_argument("delay", type=float, help="Time delay between price fetches in seconds.")
     parser.add_argument("--format", choices=['text', 'json'], default='text', help="Output format: text or json.")
     parser.add_argument("--depth", type=int, default=5, help="Number of levels to fetch from the order book.")
     args = parser.parse_args()
